requirements.txt
bottleneck-transformer-pytorch>=0.1.4
cached-property>=1.5.2
certifi>=2021.10.8
cffi>=1.14.6
charset-normalizer>=2.0.7
click>=8.0.3
Cython>=0.29.24
dataclasses>=0.6
deepsmiles>=1.0.1
dropblock>=0.3.0
efficientnet-pytorch>=0.7.1
einops>=0.3.2
future>=0.18.2
fvcore>=0.1.5.post20211023
gensim>=4.3.2
numpy>=1.26.2
h5py>=3.10.0
iopath>=0.1.9
joblib>=1.3.2
Levenshtein>=0.16.0
nltk>=3.8.1
pandas>=1.3.3
Pillow>=8.3.1
rapidfuzz>=1.8.2
regex>=2021.10.8
requests>=2.26.0
resnest>= 0.0.6b20211220
scikit-learn>=0.20.4
scipy>=1.7.0
smart-open>=5.2.1
tabulate>P=0.8.9
termcolor>=1.1.0
Theano>=1.0.5
thop>=0.0.31.post2005241907
threadpoolctl>=3.0.0
timm>=0.3.2
torch>=2.1.0
torchsummary>=1.5.1
torchvision>=0.8.2
tqdm>=4.62.3
typing-extensions>=3.10.0.2
urllib3>=1.26.7
xdg>=1.0.7
yacs>=0.1.8
zipp>=3.6.0
apex
